+++ src/routes/associations.js
@@ -41,15 +41,11 @@ module.exports = function (app, model, Implementation, integrator, opts) {
       .then(function (results) {
         var records = results[0];
         var count = results[1];
-        var decorators = results[2];
+        var fieldsSearched = results[2];
 
         var meta = {
           count: count
         };
-
-        if (decorators) {
-          meta.decorators = decorators;
-        }
 
         return new ResourceSerializer(
           Implementation,
@@ -57,7 +53,9 @@ module.exports = function (app, model, Implementation, integrator, opts) {
           records,
           integrator,
           opts,
-          meta
+          meta,
+          fieldsSearched,
+          params.search
         ).perform();
       })
       .then(function (records) { response.send(records); })