+++ src/serializers/resource.js
@@ -8,7 +8,7 @@ var Schemas = require('../generators/schemas');
 var logger = require('../services/logger');
 
 function ResourceSerializer(Implementation, model, records, integrator, opts, meta,
-  fieldsPerModel) {
+  fieldsSearched, searchValue, fieldsPerModel) {
   var modelName = Implementation.getModelName(model);
   var schema = Schemas.schemas[modelName];
 
@@ -162,12 +162,39 @@ function ResourceSerializer(Implementation, model, records, integrator, opts, me
     return new P(function (resolve) {
       if (_.isArray(records)) {
         resolve(P.map(records, function (record) {
-          return new SmartFieldsValuesInjector(record, modelName, fieldsPerModel).perform();
+          var smartFieldsValuesInjector =
+            new SmartFieldsValuesInjector(record, modelName, fieldsPerModel);
+
+          return smartFieldsValuesInjector.perform()
+            .then(function (result) {
+              fieldsSearched = fieldsSearched.concat(smartFieldsValuesInjector.getFieldsSearched());
+              return result;
+            });
         }));
       } else {
-        resolve(new SmartFieldsValuesInjector(records, modelName, fieldsPerModel).perform());
+        var smartFieldsValuesInjector =
+          new SmartFieldsValuesInjector(records, modelName, fieldsPerModel);
+
+        resolve(smartFieldsValuesInjector.perform()
+          .then(function (result) {
+            fieldsSearched = fieldsSearched.concat(smartFieldsValuesInjector.getFieldsSearched());
+            return result;
+          }));
       }
     })
+      .then(function () {
+        var decorators = null;
+        if (searchValue) {
+          decorators = Implementation.RecordsDecorator.decorateForSearch(
+            records,
+            fieldsSearched,
+            searchValue
+          );
+          if (decorators) {
+            serializationOptions.meta.decorators = decorators;
+          }
+        }
+      })
       .then(function () {
         return new JSONAPISerializer(schema.name, records, serializationOptions);
       });